Metodo ventaConMontoMasAlto

+++ tp_Final/src/main/java/com/redo/tp_Final/dto/VentaDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class VentaDTO {
-    private int codigoVenta;
+    private Long codigoVenta;
     private Double total;
     private int cantidadProductos;
     private String nombreCliente;
